fix: address pre-PR review findings on comm-channel-work
Gate AGENT-channel reads behind Execute (not Preview) via a preview flag
threaded through step generation; clamp GOVBURN's stale-tick day count to
avoid over-buying; split agent-channel.ts's DOM automation into a separate
prun-ui module; remove dead beforeExecute/openAgentChannel code; fix a
stale watcher dependency, a populationId null-out, and a few style nits.
